Increase buffer size of ReadDir to permit larger file names.
Experimentally, very lengthy file names somehow do not fit in the 256
character buffer, even though they should (NAME_MAX).  Adding 8 or 16
characters to the size of the buffer I tried, and it did not work.
So I doubled the character buffer size to be on the safe side, and
this worked.  Noted that I tried 384 bytes too, and this worked, but
it is not clear to me if (counting getdents64 offsets) that would be
safe in the presence of two files close to NAME_MAX consecutively in
a directory listing, so I chose the safe option instead.